requirements.txt
Appium-Python-Client~=4.0.0
selenium~=4.22.0
pytest~=8.2.2

Contexts~=0.12
